chatbot_api/requirements.txt
asyncio==3.4.3
chromadb==0.4.22
duckduckgo-search==6.1.6
e2b==0.17.1
fastapi-cli==0.0.4
fastapi==0.111.0
google-cloud-logging==3.10.0
jq==1.7.0
langchain-community==0.2.5
langchain-core==0.2.7
langchain-openai==0.1.9
langchain-text-splitters==0.2.1
langchain==0.2.5
langchainhub==0.1.14
lark==1.1.9
numpy==1.26.4
openai==1.34.0
opentelemetry-api==1.22.0
pydantic_core==2.18.4
pydantic==2.7.4
pymongo==4.7.3
pytest==8.2.2
python-dotenv==1.0.1
transformers==4.41.0
uvicorn==0.30.1